Add files via upload
A simple, interactive Tic-Tac-Toe game made in Python which lets two players play against each other in the terminal. Players take turns choosing positions on a 3×3 grid.
Features:

Two-player gameplay (Player X vs Player O)

Input validation for invalid or occupied spots



Automatic win and tie detection



Option to replay after each round.



How it works:

The game board is represented as a list of 9 elements (indices 0–8). For every turn, the player chooses a position by entering a number that matches to a grid spot. The program updates the board, checks for a win or tie, and switches turns accordingly.